Move some ethereum structures into the specific module they belong
