+++ rotkehlchen/chain/ethereum/modules/eth2/constants.py
@@ -1,0 +1,11 @@
+from rotkehlchen.constants.ethereum import EthereumConstants
+
+ETH2_DEPOSIT = EthereumConstants().contract('ETH2_DEPOSIT')
+
+REQUEST_DELTA_TS = 60 * 60  # 1
+
+VALIDATOR_STATS_QUERY_BACKOFF_EVERY_N_VALIDATORS = 30
+VALIDATOR_STATS_QUERY_BACKOFF_TIME_RANGE = 20
+VALIDATOR_STATS_QUERY_BACKOFF_TIME = 8
+
+FREE_VALIDATORS_LIMIT = 4
